# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eye Analytics API",
        "description": "REST API for the Eye Analytics platform. All endpoints are prefixed with `/api/v1`.\r\n *\r\n * ## Authentication\r\n * Protected endpoints require a Sanctum bearer token obtained from `POST /api/v1/auth/login`.\r\n * Pass it as: `Authorization: Bearer {token}`\r\n *\r\n * ## App-level API Keys\r\n * Every request must also carry the shared frontend key pair:\r\n * - `X-Public-Key: pk_xxxx`\r\n * - `X-Secret-Key: sk_xxxx`\r\n * Generate these once with `php artisan app:generate-api-keys`.",
        "contact": {
            "email": "support@eye-analytics.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current environment"
        }
    ],
    "paths": {
...
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user account",
                "operationId": "1bc076b4e4d21f72f56edfa37e6364c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "UTC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created — returns token"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer token — obtain via POST /api/v1/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}